# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sayohatchilar o'rtasida tajriba almashish, sayohat rejalarini tuzish va boshqalar bilan bo'lishish imkoniyatini yaratish orqali global sayohat hamjamiyatini shakllantirish.",
        "title": "TravelTales App",
        "contact": {
            "name": "Saidakbar",
            "url": "http://www.support_me_with_smile",
            "email": "\"pardaboyevsaidakbar103@gmail.com\""
        },
        "version": "1.0"
    },
    "host": "localhost:4444",
    "basePath": "/travel",
    "paths": {
...
        "/users": {
            "get": {
                "description": "this is for getting user information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get Users info",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit is required",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page is required",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns users information",
                        "schema": {
                            "$ref": "#/definitions/users.ResponseGetUsers"
                        }
                    },
                    "400": {
                        "description": "It occurs when user enter invalid params",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "It occurs when error happenes internal service",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/users/update": {
            "put": {
                "description": "This endpoint is for updating password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update Password",
                "parameters": [
                    {
                        "description": "User ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.RequestUpdatePassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message about updating password",
                        "schema": {
                            "$ref": "#/definitions/users.ResponseUpdatePassword"
                        }
                    },
                    "400": {
                        "description": "Occurs when user enters invalid params",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Occurs when an internal service error happens",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/users/{id}/GetProfile": {
            "get": {
                "description": "this is for getting a user information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get User information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id is required",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns users info",
                        "schema": {
                            "$ref": "#/definitions/users.ResponseGetProfile"
                        }
                    },
                    "400": {
                        "description": "It occurs when user enter invalid params",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "It occurs when error happenes internal service",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/users/{id}/delete": {
            "delete": {
                "description": "This endpoint is for deleting a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message about deleting user",
                        "schema": {
                            "$ref": "#/definitions/users.ResponseDeleteUser"
                        }
                    },
                    "400": {
                        "description": "Occurs when user enters invalid params",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Occurs when an internal service error happens",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/users/{id}/editprofile": {
            "post": {
                "description": "this is for editing user information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Edit Profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id is required",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "all params are required",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.RequestEditProfile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns user information about existing user",
                        "schema": {
                            "$ref": "#/definitions/users.ResponseEditProfile"
                        }
                    },
                    "400": {
                        "description": "It occurs when user enter invalid params",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "It occurs when error happenes internal service",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/users/{id}/follow": {
            "post": {
                "description": "this is for following user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Follow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id is required",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "all params are required and followoreId is taking from params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.RequestFollow"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns all ditails about following",
                        "schema": {
                            "$ref": "#/definitions/users.ResponseFollow"
                        }
                    },
                    "400": {
                        "description": "It occurs when user enter invalid params",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "It occurs when error happenes internal service",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/users/{id}/followers": {
            "get": {
                "description": "this is for getting followers information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get Followers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id is required",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit is required",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page is required",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns followers information",
                        "schema": {
                            "$ref": "#/definitions/users.ResponseGetFollowers"
                        }
                    },
                    "400": {
                        "description": "It occurs when user enter invalid params",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "It occurs when error happenes internal service",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
        "/users/{id}/validate": {
            "get": {
                "description": "this is for chacking user is exists",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Validate User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id is required",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns boolean value about existing user",
                        "schema": {
                            "$ref": "#/definitions/users.Status"
                        }
                    },
                    "400": {
                        "description": "It occurs when user enter invalid params",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "It occurs when error happenes internal service",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "users.Follower": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "users.RequestEditProfile": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "countries_visited": {
                    "type": "integer"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "users.RequestFollow": {
            "type": "object",
            "properties": {
                "follower_id": {
                    "type": "string"
                },
                "following_id": {
                    "type": "string"
                }
            }
        },
        "users.RequestUpdatePassword": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                }
            }
        },
        "users.ResponseDeleteUser": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "users.ResponseEditProfile": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "countries_visited": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "users.ResponseFollow": {
            "type": "object",
            "properties": {
                "followed_at": {
                    "type": "string"
                },
                "follower_id": {
                    "type": "string"
                },
                "following_id": {
                    "type": "string"
                }
            }
        },
        "users.ResponseGetFollowers": {
            "type": "object",
            "properties": {
                "followers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/users.Follower"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "users.ResponseGetProfile": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "countries_visited": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "users.ResponseGetUsers": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/users.User"
                    }
                }
            }
        },
        "users.ResponseUpdatePassword": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "users.Status": {
            "type": "object",
            "properties": {
                "Success": {
                    "type": "boolean"
                }
            }
        },
        "users.User": {
            "type": "object",
            "properties": {
                "countries_visited": {
                    "type": "integer"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}